requirements.txt
asgiref
blinker
click
colorama
crispy-bootstrap5
dj-database-url
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
Flask
gunicorn
itsdangerous
Jinja2
Markdown
MarkupSafe
packaging
Pillow
psycopg2-binary
pytz
sqlparse
typing_extensions
tzdata
Werkzeug
